refactor: moving the ApplicationInsights client into common
