+++ azurerm/internal/clients/client_applicationinsights.go
@@ -1,18 +1,18 @@
-package applicationinsights
+package clients
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/appinsights/mgmt/2015-05-01/insights"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/common"
 )
 
-type Client struct {
+type ApplicationInsightsClient struct {
 	AnalyticsItemsClient *insights.AnalyticsItemsClient
 	APIKeysClient        *insights.APIKeysClient
 	ComponentsClient     *insights.ComponentsClient
 	WebTestsClient       *insights.WebTestsClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func newApplicationInsightsClient(o *common.ClientOptions) *ApplicationInsightsClient {
 	analyticsItemsClient := insights.NewAnalyticsItemsClient(o.SubscriptionId)
 	o.ConfigureClient(&analyticsItemsClient.Client, o.ResourceManagerAuthorizer)
 
@@ -25,7 +25,7 @@ func BuildClient(o *common.ClientOptions) *Client {
 	webTestsClient := insights.NewWebTestsClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&webTestsClient.Client, o.ResourceManagerAuthorizer)
 
-	return &Client{
+	return &ApplicationInsightsClient{
 		AnalyticsItemsClient: &analyticsItemsClient,
 		APIKeysClient:        &apiKeysClient,
 		ComponentsClient:     &componentsClient,
